+++ app/lib/account/models.dart
@@ -49,14 +49,6 @@ class User extends db.ExpandoModel<String> {
   @db.BoolProperty(required: true)
   bool isDeleted = false;
 
-  /// [isBlocked] is set when a user account is blocked (is on administrative hold).
-  /// When this happens user-data is preserved, but the user should not be able
-  /// to perform any action.
-  ///
-  /// TODO: remove after runtime version `2024.12.17` is no longer running.
-  @db.BoolProperty(required: false)
-  bool isBlocked = false;
-
   /// `true` if user was moderated (pending moderation or deletion).
   @db.BoolProperty(required: true)
   bool isModerated = false;